+++ gtime/feature_extraction/tests/test_sorted_density.py
@@ -3,7 +3,7 @@ import pandas as pd
 import pandas.util.testing as testing
 import pytest
 
-from gtime.custom.crest_factor_detrending import CrestFactorDetrending
+from gtime.feature_extraction.custom import SortedDensity
 
 
 def get_input_data():
@@ -17,12 +17,12 @@ def get_input_data():
 
 
 def get_output_causal():
-    custom_feature = CrestFactorDetrending(window_size=2, is_causal=True)
+    custom_feature = SortedDensity(window_size=2, is_causal=True)
     feature_name = custom_feature.__class__.__name__
     output_causal = pd.DataFrame.from_dict(
         {
-            f"x_1__{feature_name}": [np.nan, 1.0, 0.07547169811320754],
-            f"x_2__{feature_name}": [np.nan, 0.9615384615384616, 0.13793103448275862],
+            f"x_1__{feature_name}": [np.nan, 0.5, 0.6111111111111112],
+            f"x_2__{feature_name}": [np.nan, 0.5833333333333334, 0.6428571428571429],
         }
     )
     output_causal.index = [
@@ -34,12 +34,12 @@ def get_output_causal():
 
 
 def get_output_anticausal():
-    custom_feature = CrestFactorDetrending(window_size=2, is_causal=False)
+    custom_feature = SortedDensity(window_size=2, is_causal=False)
     feature_name = custom_feature.__class__.__name__
     output_anticausal = pd.DataFrame.from_dict(
         {
-            f"x_1__{feature_name}": [1.0, 0.07547169811320754],
-            f"x_2__{feature_name}": [0.9615384615384616, 0.13793103448275862],
+            f"x_1__{feature_name}": [0.5, 0.6111111111111112],
+            f"x_2__{feature_name}": [0.5833333333333334, 0.6428571428571429],
         }
     )
     output_anticausal.index = [
@@ -54,15 +54,15 @@ output_causal = get_output_causal()
 output_anticausal = get_output_anticausal()
 
 
-class TestCrestFactorDetrending:
+class TestSortedDensity:
     @pytest.mark.parametrize("test_input, expected", [(input_data, output_causal)])
     def test_crest_factor_detrending_causal(self, test_input, expected):
-        feature = CrestFactorDetrending(window_size=2, is_causal=True)
+        feature = SortedDensity(window_size=2, is_causal=True)
         output = feature.fit_transform(test_input)
         testing.assert_frame_equal(output, expected)
 
     @pytest.mark.parametrize("test_input, expected", [(input_data, output_anticausal)])
     def test_crest_factor_detrending_anticausal(self, test_input, expected):
-        feature = CrestFactorDetrending(window_size=2, is_causal=False)
+        feature = SortedDensity(window_size=2, is_causal=False)
         output = feature.fit_transform(test_input)
         testing.assert_frame_equal(output, expected)
